makes sure intervals or ints

+++ teleceptor/timeAggregationUtils.py
@@ -64,14 +64,14 @@ def getElasticSearchAggregationLevel(start, end):
     """
     aggregation_period = getAggregationLevel(start/1000, end/1000)
     if aggregation_period < 60:
-        return "{}s".format(aggregation_period)
+        return "{}s".format(int(aggregation_period))
     if aggregation_period < 60*60:
         # One hour
-        return "{}m".format(aggregation_period/60.0)
+        return "{}m".format(int(aggregation_period/60.0))
     if aggregation_period < 60*60*24:
         # One day
-        return "{}h".format((aggregation_period/60.0)/60.0)
-    return "{}d".format(((aggregation_period/60.0)/60.0)/24.0)
+        return "{}h".format(int((aggregation_period/60.0)/60.0))
+    return "{}d".format(int(((aggregation_period/60.0)/60.0)/24.0))
 
 
 if __name__ == "__main__":
